STYLE: Basic paragraph styling
Added text to Goataclause page to test with
Added styling for <p> and <hr> tags to begin the de-uglifying of page content.
Added color styling for links
Fixed footer links to be compatible with new colors
Updated utility class for custom alt text to be font 0
Fixed typo in footer copy text.